Fixed permissions in sort_by_key, added missing cleanup, and enabled to work with python 3 (as well as 2).
Former-commit-id: 485150ff8c1d448221bdfa2af17bb9b9216049c7

+++ python/examples/sort_by_key/src/sort.py
@@ -37,7 +37,7 @@ def chunks(l, n, balanced=False):
 
 def partition_by(self, num_partitions, hash):
     if not hash:
-        return chunks(self, len(self) / num_partitions, True)
+        return chunks(self, int(len(self) / num_partitions), True)
     else:
         partitions = [[] for n in range(num_partitions)]
         for s in self:
